verify_locks_script.py
import requests
import json

# Login to get token (simulated or just check public endpoint if auth not required for this check, 
# but we need to check for a free user, so we need a user context)
# Actually, the view uses request.user.
# I'll use the existing user 'anujyadavvvv12345@gmail.com' who should be free unless I made him premium.

# First, let's just check the logic by running a django shell script instead of a full HTTP request 
# because authentication might be tricky to script quickly without a token.
# I will run this via 'python manage.py shell'
print("This file is intended to be run via django shell, not directly.")
