requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
firebase-admin==6.3.0
razorpay==1.4.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
